+++ src/main/java/entity/User.java
@@ -1,4 +1,4 @@
-package Entity;
+package entity;
 
 import java.sql.Timestamp;
 import org.springframework.cloud.gcp.data.spanner.core.mapping.PrimaryKey;
