+++ src/main/java/exceptions/InvalidLoginException.java
@@ -1,16 +1,15 @@
-package Exceptions;
+package exceptions;
 
 import java.lang.RuntimeException;
 import org.springframework.http.HttpStatus;
-import Exceptions.APIException;
 
-public class UserNotFoundException extends RuntimeException implements APIException {
+public class InvalidLoginException extends RuntimeException implements APIException {
 
     private static final HttpStatus HTTP_STATUS = HttpStatus.NOT_FOUND;
-    private static final String MESSAGE = "Invalid Username";
+    private static final String MESSAGE = "Invalid Username or Password";
     private final String path;
 
-    public UserNotFoundException(String path) {
+    public InvalidLoginException(String path) {
         super();
         this.path = path;
     }
