+++ src/main/java/exceptions/UserNotFoundException.java
@@ -1,16 +1,15 @@
-package Exceptions;
+package exceptions;
 
 import java.lang.RuntimeException;
 import org.springframework.http.HttpStatus;
-import Exceptions.APIException;
 
-public class UserAlreadyExistsException extends RuntimeException implements APIException {
+public class UserNotFoundException extends RuntimeException implements APIException {
 
-    private static final HttpStatus HTTP_STATUS = HttpStatus.CONFLICT;
-    private static final String MESSAGE = "Username or Email-ID already exists";
+    private static final HttpStatus HTTP_STATUS = HttpStatus.NOT_FOUND;
+    private static final String MESSAGE = "Invalid Username";
     private final String path;
 
-    public UserAlreadyExistsException(String path) {
+    public UserNotFoundException(String path) {
         super();
         this.path = path;
     }
