+++ src/test/java/main/TestClass.java
@@ -1,4 +1,4 @@
-package Main;
+package main;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
